refactor: move AuthenticationService and rename exception
