+++ admin_panel/src/commonMain/kotlin/net/thechance/mena/admin_panel/domain/exceptions/AdminPanelExceptions.kt
@@ -9,5 +9,5 @@ class UnknownNetworkException(message: String = "Unknown error") : AdminPanelExc
 class InvalidPasswordException(message: String = "password doesn't match validations") :
     AdminPanelException(message)
 
-class InvalidCredentialsException(message: String = "Invalid user name or password") :
+class UnauthorizedException(message: String = "Invalid user name or password") :
     AdminPanelException(message)
